% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{forecasthistogram_example}
\alias{forecasthistogram_example}
\title{Draw example histogram}
\usage{
forecasthistogram_example(
  style = "Gaussian",
  n_bins = NULL,
  enforce_adjacent = TRUE
)
}
\arguments{
\item{style}{Either "Gaussian" or "Random". "Gaussian" means that histogram has (approximate) bell shape.}

\item{n_bins}{Number of bins that receive strictly positive probability. Defaults to \code{NULL} (random number of bins). Relevant only if \code{style == "Random"}.}

\item{enforce_adjacent}{Whether to enforce adjacency of bins (defaults to \code{TRUE}). Relevant only if \code{style == "Random"}.}
}
\description{
Draw example histogram
}
\details{
Bin definitions of random histogram are as in the Survey of Consumer Expectations data set.
}
\examples{
# Draw random histogram
x <- forecasthistogram_example()
# Plot it
plot(x)
}
