% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{mean.forecasthistogram}
\alias{mean.forecasthistogram}
\title{Compute mean of fitted histogram}
\usage{
\method{mean}{forecasthistogram}(x, ...)
}
\arguments{
\item{x}{object of class \code{forecasthistogram}}

\item{...}{additional parameters, currently not in use}
}
\description{
Compute mean of fitted histogram
}
