% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{num_to_bin}
\alias{num_to_bin}
\title{Helper function to convert numerical realization into bin indicator}
\usage{
num_to_bin(y, ub)
}
\arguments{
\item{y}{numerical value}

\item{ub}{vector of upper bounds of histogram bin ranges}
}
\description{
Helper function to convert numerical realization into bin indicator
}
