% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{plot.forecasthistogram}
\alias{plot.forecasthistogram}
\title{Function to visualize SCE histogram forecast using ggplot2}
\usage{
\method{plot}{forecasthistogram}(x, ylim = NULL, outer_width = 3, ...)
}
\arguments{
\item{x}{object of class `forecasthistogram'}

\item{ylim}{support limit for vertical axis (defaults to \code{NULL}, useful to make several comparable plots)}

\item{outer_width}{support extension in case of infinite limits}

\item{...}{additional parameters (currently not in use)}
}
\value{
plot (ggplot2 object)
}
\description{
Function to visualize SCE histogram forecast using ggplot2
}
