% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{quantile.forecasthistogram}
\alias{quantile.forecasthistogram}
\title{Compute quantiles of fitted histogram}
\usage{
\method{quantile}{forecasthistogram}(x, probs = (1:9)/10, ...)
}
\arguments{
\item{x}{object of class \code{forecasthistogram}}

\item{probs}{quantile levels}

\item{...}{additional parameters, currently not in use}
}
\description{
Compute quantiles of fitted histogram
}
