% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{rps_tools}
\alias{rps_tools}
\alias{rps}
\alias{erps}
\alias{ebs}
\alias{erps.numeric}
\alias{erps.forecasthistogram}
\alias{ebs.numeric}
\alias{ebs.forecasthistogram}
\alias{rps.numeric}
\alias{rps.forecasthistogram}
\title{Functions for the (Expected) Ranked Probability Score}
\usage{
\method{erps}{numeric}(x)

\method{erps}{forecasthistogram}(x)

\method{ebs}{numeric}(x)

\method{ebs}{forecasthistogram}(x)

\method{rps}{numeric}(x, kstar)

\method{rps}{forecasthistogram}(x, kstar)
}
\arguments{
\item{x}{object of class `forecasthistogram', or vector of probabilities (must be positive and sum to one)}

\item{kstar, }{realizing category (integer, min = 1, max = length(p))}
}
\value{
\code{erps} returns the Expected Ranked Probability Score (ERPS). \code{rps} returns the Ranked Probability Score (RPS). \code{ebs} returns the expected Brier score (EBS).
}
\description{
Functions for the (Expected) Ranked Probability Score
}
\details{
Krüger and Pavlova (2020) propose to use the ERPS as a robust measure
of the uncertainty in a survey histogram. The RPS is due to Epstein (1969).
See also Section 4.2 in Krüger and Pavlova (2020).
}
\references{
Epstein, E.S. (1969): `A Scoring System for Probability Forecasts of Ranked Categories', Journal of Applied Meteorology and Climatology 8, 985-987.

Krüger, F., and L. Pavlova (2020): `Quantifying Subjective
Uncertainty in Survey Expectations'. Working Paper, \doi{10.18452/21401}.
}
