% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{triangular}
\alias{triangular}
\alias{dtri}
\alias{ptri}
\alias{moments_tri}
\title{Density and CDF of triangular distribution}
\usage{
dtri(x, a, b, c)

ptri(x, a, b, c)

moments_tri(a, b, c)
}
\arguments{
\item{x}{evaluation point for density or CDF}

\item{a, b, c}{parameters of distribution (left limit, center, right limit)}
}
\value{
\code{dtri} returns the density, \code{ptri} returns the CDF,
\code{moments_tri} returns the mean and standard deviation
}
\description{
Density and CDF of triangular distribution
}
