% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{var.forecasthistogram}
\alias{var.forecasthistogram}
\title{Compute variance of fitted histogram}
\usage{
var.forecasthistogram(x, ...)
}
\arguments{
\item{x}{object of class \code{forecasthistogram}}

\item{...}{additional parameters, currently not in use}
}
\description{
Compute variance of fitted histogram
}
