% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_tools.R
\name{forecasthistogram}
\alias{forecasthistogram}
\title{Constructor function for forecasthistogram object}
\usage{
forecasthistogram(ub, p)
}
\arguments{
\item{ub}{vector of upper bounds}

\item{p}{vector of probabilities}
}
\description{
Constructor function for forecasthistogram object
}
